Extract StepGraph struct to src/qwen35/step_graph.h
Move StepGraph struct and step_graph_free/step_graph_destroy utility
functions from test_dflash.cpp into a reusable header. test_dflash.cpp
now includes step_graph.h with using-declarations.
